gigaq: VRED: STILL: BUILD

+++ autopath/models/vred.py
@@ -481,7 +481,8 @@ class VariationalReDecoderStill(Datablock):
         # Link the logs directory to the provided location (e.g., for Tensorboard to pick up the logs)
         if self.logs is not None:
             self.log.verbose(f"Linking logs to {self.logs}")
-            os.unlink(self.logs)
+            if os.path.exists(self.logs):
+                os.unlink(self.logs)
             os.path.symlink(self.dirpath('logs'), self.logs)
 
     def valid(self):
